refactor: separated ProposalHistory in its own component
