+++ src/Components/ProposalInfoCard/ProposalHistory/ProposalHistory.types.tsx
@@ -1,0 +1,5 @@
+import { Proposal } from 'types/types.guilds.d';
+
+export interface ProposalHistoryProps {
+  proposal: Proposal;
+}
