+++ src/Components/ProposalInfoCard/ProposalHistory/index.ts
@@ -1,0 +1,2 @@
+export { default as ProposalHistory } from './ProposalHistory';
+
